Implement rough calculator functionality
The calculator works reasonably well and correctly displays input
digits and performs the correct operations on them. It also correctly
displays the result.

However, there is currently an issue where if you try to enter
a new number after a successful calculation and then an operator
and a second number, the calculator will always use the previous result
instead of the new first number. This should be a relatively easy fix
with the use of a flag variable or some reworking of the logic.

Also need to refactor some repetitive code into functions, particularly
the code involved when the operator and equals button are pressed.